Upgrade to Laravel 5.5
This is a retrofit of the 5.1-based application on top of a clean
Laravel installation. The application structure is more-or-less the
same, but uses more current Laravel norms. The use of Vue has also
been refined by refactoring to single-file components and putting more
display logic on the client-side where it can be done less verbosely
than it was before. The UI has also been reworked around Bootstrap tab
navigation.