Rework pickable component to include form input
Rather than have the component remotely populate the value of the text
field it is associated with, make the text field part of the
component. This allows for more reactivity, and takes better advantage
of Vue. Now the shortcuts are able to reflect the current value. The
view partial is also able to specify which shortcut groups it wants,
since the shortcuts for time and invoice entry deal with dates at
different scales.