Expose project time budget fields on project.show and project.form
The Project model exposes the newly-added time budget fields both
directly as well as in hourly form via custom attributes. Additional
computed accessors are used to calculate time remaining.

The Time::forProjectAndUserByInterval() is generalized here because it
was initially going to be used for the time remaining calculations,
but ended up not. The generalization is otherwise useful though.