Consolidate forceDelete in console command
Having each model be responsible for pruning via trait was annoying
due to things being spread around in multiple places, and due to
eloquent events not being fired unless the deletion was done in a
particular way. There is also only an after-the-fact event for force
deletes, which is not useful when dealing with foreign key constraints.