Store user timezone during login and registration.
The previous approach didn't account for the scenario where local
times need to be displayed at the first page load prior to the
timezone session cookie being available, such as when viewing the
dashboard with in-progress time entries. Instead of relying on the
cookie, store the value in the user table. This impacts the
TimeHelper, where the centralizing timezone conversion in a dedicated
method was no longer convenient. It also impacts how time and invoice
suggestions are produced. The calculated attribute suggestion is now a
method so that the timezone can be specified.