Return to prefixed invoice numbers.
Number invoices by concatenating the project id with the current tally
of invoices for the project. This is similar to the previous approach
of prefixing by client id, but easier to assemble. Just incrementing
from the previous number without prefixing was ok when starting from
scratch, but thrown off by client id approach.